app-defaults: refactor things into folders and move things closer to home
