+++ packages/app-defaults/src/createApp/createApp.test.tsx
@@ -18,84 +18,7 @@ import { screen } from '@testing-library/react';
 import { renderWithEffects } from '@backstage/test-utils';
 import React, { PropsWithChildren } from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import { defaultConfigLoader, createApp } from './createApp';
-
-(process as any).env = { NODE_ENV: 'test' };
-const anyEnv = process.env as any;
-const anyWindow = window as any;
-
-describe('defaultConfigLoader', () => {
-  afterEach(() => {
-    delete anyEnv.APP_CONFIG;
-    delete anyWindow.__APP_CONFIG__;
-  });
-
-  it('loads static config', async () => {
-    anyEnv.APP_CONFIG = [
-      { data: { my: 'config' }, context: 'a' },
-      { data: { my: 'override-config' }, context: 'b' },
-    ];
-
-    const configs = await defaultConfigLoader();
-    expect(configs).toEqual([
-      { data: { my: 'config' }, context: 'a' },
-      { data: { my: 'override-config' }, context: 'b' },
-    ]);
-  });
-
-  it('loads runtime config', async () => {
-    anyEnv.APP_CONFIG = [
-      { data: { my: 'override-config' }, context: 'a' },
-      { data: { my: 'config' }, context: 'b' },
-    ];
-
-    const configs = await (defaultConfigLoader as any)(
-      '{"my":"runtime-config"}',
-    );
-    expect(configs).toEqual([
-      { data: { my: 'override-config' }, context: 'a' },
-      { data: { my: 'config' }, context: 'b' },
-      { data: { my: 'runtime-config' }, context: 'env' },
-    ]);
-  });
-
-  it('fails to load invalid missing config', async () => {
-    await expect(defaultConfigLoader()).rejects.toThrow(
-      'No static configuration provided',
-    );
-  });
-
-  it('fails to load invalid static config', async () => {
-    anyEnv.APP_CONFIG = { my: 'invalid-config' };
-    await expect(defaultConfigLoader()).rejects.toThrow(
-      'Static configuration has invalid format',
-    );
-  });
-
-  it('fails to load bad runtime config', async () => {
-    anyEnv.APP_CONFIG = [{ data: { my: 'config' }, context: 'a' }];
-
-    await expect((defaultConfigLoader as any)('}')).rejects.toThrow(
-      'Failed to load runtime configuration, SyntaxError: Unexpected token } in JSON at position 0',
-    );
-  });
-
-  it('loads config from window.__APP_CONFIG__', async () => {
-    anyEnv.APP_CONFIG = [
-      { data: { my: 'config' }, context: 'a' },
-      { data: { my: 'override-config' }, context: 'b' },
-    ];
-    const windowConfig = { app: { configKey: 'config-value' } };
-    anyWindow.__APP_CONFIG__ = windowConfig;
-
-    const configs = await defaultConfigLoader();
-
-    expect(configs).toEqual([
-      ...anyEnv.APP_CONFIG,
-      { context: 'window', data: windowConfig },
-    ]);
-  });
-});
+import { createApp } from './createApp';
 
 describe('Optional ThemeProvider', () => {
   it('should render app with user-provided ThemeProvider', async () => {
