+++ packages/app-defaults/src/createApp/defaults/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2021 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,4 +14,8 @@
  * limitations under the License.
  */
 
-export { createApp } from './createApp';
+export { apis } from './apis';
+export { components } from './components';
+export { configLoader } from './configLoader';
+export { icons } from './icons';
+export { themes } from './themes';
